refactor: filter factory
